more modifications

+++ src/main/java/com/zone/innovation/activityfeed/dao/FeedRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.zone.innovation.activityfeed.model.Feed;
 
 @Repository
-public interface FeedRepository extends CrudRepository<Feed,Integer>{
+public interface FeedRepository extends CrudRepository<Feed,Long>{
 
 }
